requirements.txt
# ============================================================
# AI ROOT HEALTH MONITOR — Python Dependencies
# ============================================================
# Python 3.10+ REQUIRED (NOT 3.13)

# ── PYTORCH & DEEP LEARNING ──────────────────────────────
# Install torch with GPU support or CPU separately:
# GPU:  pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# CPU:  pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ── DATA PROCESSING & ARRAYS ─────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=9.5.0

# ── IMAGE AUGMENTATION ───────────────────────────────────
albumentations>=1.3.0

# ── MACHINE LEARNING & METRICS ──────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0

# ── VISUALIZATION & MONITORING ──────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# ── FASTAPI BACKEND ─────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# ── DATABASE & ORM ──────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.12.0  # database migrations (if needed)

# ── DATA VALIDATION ─────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── ENVIRONMENT & CONFIGURATION ─────────────────────────
python-dotenv>=1.0.0

# ── DEVELOPMENT & TESTING ──────────────────────────────
httpx>=0.25.0  # for testing API endpoints
pytest>=7.4.0  # for unit tests
pytest-asyncio>=0.21.0  # for async test support

# ── UTILITIES ────────────────────────────────────────────
requests>=2.31.0  # HTTP library
python-json-logger>=2.0.0  # structured logging (optional)
